core: move browser message store to separate file and re-export from index.ts
